Add files via upload
Code for cleaning the dataset. Fergal will add a final bit of code to this. Please check the code for errors before using it, and let the team know if you make changes to the code to ensure we are all working off the same clean dataset moving forward. Thanks.